% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance_runners.R
\name{advance_runners}
\alias{advance_runners}
\title{Advance Runners on Base}
\usage{
advance_runners(result, bases)
}
\arguments{
\item{result}{the result of a plate appearance}

\item{bases}{the current base state}
}
\description{
This function is used within the sim_inning function to advance runners
}
\examples{
advance_runners(result = "Single", bases = "0,1,0")
}
\keyword{baseball}
