% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_profiles}
\alias{avg_profiles}
\title{Average Batter Profiles}
\format{
\subsection{\code{avg_profiles}}{

A data frame with 3 rows and 8 columns.
\describe{
\item{bathand}{handedness of batter}
\item{bat_kpct}{proportion of plate appearances resulting in strikeout}
\item{bat_1Bpct}{proportion of plate appearances resulting in single}
\item{bat_2Bpct}{proportion of plate appearances resulting in double}
\item{bat_3Bpct}{proportion of plate appearances resulting in triple}
\item{bat_HRpct}{proportion of plate appearances resulting in home run}
\item{bat_fbpct}{proportion of plate appearances resulting in a walk or hbp}
\item{bat_ipoutpct}{proportion of plate appearances resulting in an in play, out}
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
avg_profiles
}
\description{
Average results for batters with less than 150 plate appearances in 2024.
}
\keyword{datasets}
