% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_regress.R
\name{bat_regress}
\alias{bat_regress}
\title{Regress a Hitter to the Mean}
\usage{
bat_regress(player, profiles = avg_profiles, pa_cut = 150)
}
\arguments{
\item{player}{a hitter as a data frame}

\item{profiles}{profiles for each type of handedness}

\item{pa_cut}{plate appearance cut off}
}
\description{
This function is used within the sim_pa function to give appropriate data to players under the pa cutoff
}
\examples{
bat_regress(player, profiles = avg_profiles, pa_cut = 150)
bat_regress()
}
\keyword{baseball}
