% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batters}
\alias{batters}
\title{2024 MLB Batters}
\format{
\subsection{\code{batters}}{

A data frame with 651 rows and 14 columns.
\describe{
\item{batter}{a code corresponding to the name of the batter}
\item{bathand}{the batters' handedness}
\item{bat_kpct}{the batters' strikeout percentage}
\item{bat_fbpct}{the batters' walk percentage}
\item{bat_1Bpct}{the batters' single percentage}
\item{bat_2Bpct}{the batters' double percentage}
\item{bat_3Bpct}{the batters' triple percentage}
\item{bat_HRpct}{the batters' home run percentage}
\item{bat_ipoutpct}{the batters' in play, out percentage}
\item{bat_pa}{the batters' number of plate appearances}
\item{last}{the batters' last name}
\item{first}{the batters' first name}
\item{nickname}{the batters' nickname}
\item{name}{the batters' full name}
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
batters
}
\description{
A compilation of each batter from the 2024 MLB season and their given percentages
}
\keyword{datasets}
