% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullpens}
\alias{bullpens}
\title{Default Bullpens}
\format{
\subsection{\code{bullpens}}{

A data frame with 3 rows and 30 columns.
\describe{
\item{Phillies}{bullpen of 3 Phillies pitchers}
\item{Braves}{bullpen of 3 Braves hitters}
\item{Nationals}{bullpen of 3 Nationals hitters}
...
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
bullpens
}
\description{
Default bullpens for each of the 30 MLB teams.
}
\keyword{datasets}
