% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineups}
\alias{lineups}
\title{Default Lineups}
\format{
\subsection{\code{lineups}}{

A data frame with 9 rows and 30 columns.
\describe{
\item{Phillies}{lineup of 9 Phillies hitters}
\item{Braves}{lineup of 9 Braves hitters}
\item{Nationals}{lineup of 9 Nationals hitters}
...
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
lineups
}
\description{
Default lineups for each of the 30 MLB teams.
}
\keyword{datasets}
