% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_coef}
\alias{model_coef}
\title{Model Coefficients}
\format{
\subsection{\code{model_coef}}{

A data frame with 42 rows and 3 columns.
\describe{
\item{result}{the outcome of a plate appearance}
\item{names}{the names of the coefficients}
\item{coefs}{the numerical values of the coefficients}
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
model_coef
}
\description{
A compilation of coefficients for each model from the 2024 season.
}
\keyword{datasets}
