% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_pit.R
\name{num_pit}
\alias{num_pit}
\title{Generate Number of Pitches in a Plate Appearance}
\usage{
num_pit(outcome = "Strikeout", table = pitch_table)
}
\arguments{
\item{outcome}{the result of a plate appearance}

\item{table}{a table of likelihoods that a certain number of pitches will be thrown in a pa}
}
\description{
This function is used within the sim_inning function generate the number of pitches in a plate appearance
}
\examples{
num_pit(outcome = "Double", table = pitch_table)
}
\keyword{baseball}
