% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitch_table}
\alias{pitch_table}
\title{Pitch Table}
\format{
\subsection{\code{pitch_table}}{

A data frame with 80 rows and 4 columns.
\describe{
\item{nump}{the number of pitches in a plate appearance}
\item{pit}{the amount of times this frequency of pitches was thrown}
\item{prop}{the proportion of pitch frequency}
\item{result}{the result of the plate appearance}
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
pitch_table
}
\description{
A data frame of the likelihoods that a pa will result in a certain number of pitches.
}
\keyword{datasets}
