% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pitchers}
\alias{pitchers}
\title{2024 MLB Pitchers}
\format{
\subsection{\code{pitchers}}{

A data frame with 855 rows and 14 columns.
\describe{
\item{pitcher}{a code corresponding to the name of the pitcher}
\item{pithand}{the pitchers' handedness}
\item{pit_kpct}{the pitchers' strikeout percentage}
\item{pit_fbpct}{the pitchers' walk percentage}
\item{pit_1Bpct}{the pitchers' single percentage}
\item{pit_2Bpct}{the pitchers' double percentage}
\item{pit_3Bpct}{the pitchers' triple percentage}
\item{pit_HRpct}{the pitchers' home run percentage}
\item{pit_ipoutpct}{the pitchers' in play, out percentage}
\item{pit_pa}{the pitchers' number of plate appearances}
\item{last}{the pitchers' last name}
\item{first}{the pitchers' first name}
\item{nickname}{the pitchers' nickname}
\item{name}{the pitchers' full name}
}
}
}
\source{
\url{https://retrosheet.org/downloads/plays.html}
}
\usage{
pitchers
}
\description{
A compilation of each pitcher from the 2024 MLB season and their given percentages
}
\keyword{datasets}
