% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_full_game.R
\name{sim_full_game}
\alias{sim_full_game}
\title{Simulate a Full MLB Game}
\usage{
sim_full_game(
  home_lineup = lineups$Cubs,
  home_starter = "Justin Steele",
  home_bullpen = c("Drew Smyly", "Tyson Miller", "Porter Hodge"),
  away_lineup = lineups$Twins,
  away_starter = "Joe Ryan",
  away_bullpen = c("Griffin Jax", "Cole Sands", "Jhoan Duran"),
  box_score = TRUE,
  print = "none"
)
}
\arguments{
\item{home_lineup}{lineup for the home team}

\item{home_starter}{starter for the home team}

\item{home_bullpen}{vector of bullpen pitchers for the home team}

\item{away_lineup}{lineup for the away team}

\item{away_starter}{starter for the away team}

\item{away_bullpen}{vector of bullpen pitchers for the away team}

\item{box_score}{whether to print the box_score: TRUE or FALSE}

\item{print}{whether to print the result of the pa: "none" (the default) or "result"}
}
\description{
This function lets you simulate an MLB game.
}
\examples{
sim_full_game(home_lineup=lineups$Reds,  home_starter = "Hunter Greene",
home_bullpen = c("Fernando Cruz", "Buck Farmer", "Alexis Diaz"),
away_lineup = lineups$Brewers, away_starter = "Freddy Peralta",
away_bullpen = c("Jared Koenig", "Hoby Milner", "Trevor Megill"), box_score = TRUE,
print = "none")
sim_full_game()
}
\keyword{baseball}
