% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_game.R
\name{sim_game}
\alias{sim_game}
\title{Simulates a MLB game}
\usage{
sim_game(
  lineup = lineups$Cubs,
  starter = "Sonny Gray",
  bullpen = c("Edwin Diaz", "Kenley Jansen", "Josh Hader"),
  box_score = TRUE,
  playByPlay = FALSE,
  print = "none"
)
}
\arguments{
\item{lineup}{a lineup of hitters}

\item{starter}{the starting pitcher}

\item{bullpen}{pitchers that will enter after the starter}

\item{box_score}{whether to display the box_score}

\item{playByPlay}{whether to display the play by play}

\item{print}{option to print the results of the plate appearances}
}
\description{
This function is used to simulate a MLB game
}
\examples{
sim_game(lineup = lineups$Yankees, starter = "Shane Bieber", bullpen = c("Robert Suarez", "Jeremiah Estrada", "Mason Miller"), box_score = TRUE, playByPlay = FALSE, print = "result")
sim_game()
}
\keyword{baseball}
