% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inning.R
\name{sim_inning}
\alias{sim_inning}
\title{Simulate an Inning}
\usage{
sim_inning(
  batters = "Aaron Judge",
  pit = "Paul Skenes",
  spot = 1,
  print = "none"
)
}
\arguments{
\item{batters}{list of hitters}

\item{pit}{pitcher for the inning}
}
\description{
This function lets you simulate an MLB inning.
}
\examples{
sim_inning(batters="Javier Baez", pit = "Michael Wacha")
sim_inning()
}
\keyword{baseball}
