% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pa.R
\name{sim_pa}
\alias{sim_pa}
\title{Simulate a Plate Appearance}
\usage{
sim_pa(
  bat = "Jarren Duran",
  pit = "Griffin Canning",
  pit_count = 1,
  bat_data = batters,
  pit_data = pitchers,
  print = "none"
)
}
\arguments{
\item{bat}{hitter for the pa}

\item{pit}{pitcher for the pa}

\item{pit_count}{pitch count}

\item{bat_data}{batter data}

\item{pit_data}{pitcher data}

\item{print}{whether to print the result of the pa: "none" (the default) or "result"}
}
\description{
This function lets you simulate an MLB plate appearance.
}
\examples{
sim_pa(bat="Aaron Judge", pit = "Jack Flaherty")
sim_pa()
}
\keyword{baseball}
